unsigned int/long -> size_t

+++ Engine_Graphics/HeightmapMesh.cpp
@@ -21,7 +21,7 @@ namespace Graphics
    * Index generation was adapted from here:
    * http://www.chadvernon.com/blog/resources/directx9/terrain-generation-with-a-heightmap
    */
-  HeightmapMesh::HeightmapMesh(unsigned long widthSteps, unsigned long depthSteps, float width, float depth)
+  HeightmapMesh::HeightmapMesh(size_t widthSteps, size_t depthSteps, float width, float depth)
       : m_widthSteps(widthSteps)
       , m_depthSteps(depthSteps)
   {
@@ -47,10 +47,10 @@ namespace Graphics
     const float deltaRow = depth / (float)(m_depthSteps - 1);
 
     // Generate vertices
-    unsigned int idx = 0;
-    for (unsigned int col = 0; col < m_widthSteps; col++)
+    size_t idx = 0;
+    for (size_t col = 0; col < m_widthSteps; col++)
     {
-      for (unsigned int row = 0; row < m_depthSteps; row++)
+      for (size_t row = 0; row < m_depthSteps; row++)
       {
         // Positional coordinates
         float x = (float)col * deltaCol;
@@ -72,7 +72,7 @@ namespace Graphics
 
     // Generate indices
     int index = 0;
-    for (unsigned int z = 0; z < depthSteps - 1; z++)
+    for (size_t z = 0; z < depthSteps - 1; z++)
     {
       int x;
 
@@ -81,26 +81,26 @@ namespace Graphics
       {
         for (x = 0; x < widthSteps; x++)
         {
-          m_indices[index++] = x + (z * widthSteps);
-          m_indices[index++] = x + (z * widthSteps) + widthSteps;
+          m_indices[index++] = (GLuint)(x + (z * widthSteps));
+          m_indices[index++] = (GLuint)(x + (z * widthSteps) + widthSteps);
         }
 
         // Degenerate triangle at end of non-last row
         if (z != depthSteps - 2)
-          m_indices[index++] = --x + (z * widthSteps);
+          m_indices[index++] = (GLuint)(--x + (z * widthSteps));
       }
       // Odd row (right to left)
       else
       {
         for (x = widthSteps - 1; x >= 0; x--)
         {
-          m_indices[index++] = x + (z * widthSteps);
-          m_indices[index++] = x + (z * widthSteps) + widthSteps;
+          m_indices[index++] = (GLuint)(x + (z * widthSteps));
+          m_indices[index++] = (GLuint)(x + (z * widthSteps) + widthSteps);
         }
 
         // Degenerate triangle at end of non-last row
         if (z != depthSteps - 2)
-          m_indices[index++] = ++x + (z * widthSteps);
+          m_indices[index++] = (GLuint)(++x + (z * widthSteps));
       }
     }
 
@@ -111,7 +111,7 @@ namespace Graphics
   {
   }
 
-  void HeightmapMesh::setHeight(unsigned int x, unsigned int y, float height, bool buffer)
+  void HeightmapMesh::setHeight(size_t x, size_t y, float height, bool buffer)
   {
     vertexPosition(x, y)[1] = height;
 
@@ -123,16 +123,16 @@ namespace Graphics
   void HeightmapMesh::setHeight(float *height)
   {
     // Update y coordinates
-    for (unsigned int i = 0; i < m_numVertices; i++)
+    for (size_t i = 0; i < m_numVertices; i++)
       m_vertices[i][1] = height[i];
 
     // Update buffers
     bufferData();
   }
 
-  Engine::Maths::Vector3 &HeightmapMesh::vertexPosition(unsigned int x, unsigned int y)
+  Engine::Maths::Vector3 &HeightmapMesh::vertexPosition(size_t x, size_t y)
   {
-    unsigned long idx = (x * m_depthSteps) + y;
+    size_t idx = (x * m_depthSteps) + y;
     return m_vertices[idx];
   }
 }
